+++ debitaccount.h
@@ -8,7 +8,7 @@ class DebitAccount: public CreditCard
 public:
     DebitAccount();
 
-    DebitAccount(int cardNum, int shabaNum, int accountNum, int cvv, float balance,int expireYear, int expireMonth, QString password, QString secondPassword);
+    DebitAccount(long long cardNum, long long shabaNum, long long accountNum, int cvv, double balance,int expireYear, int expireMonth, QString password, QString secondPassword);
 
     void moneyTransfer() override;
     QString toFileString() const override;
